% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visunet.R
\docType{package}
\name{VisuNet-package}
\alias{VisuNet}
\alias{VisuNet-package}
\title{VisuNet}
\description{
This is a mostly empty package to learn roxygen documentation.
}
\details{
Hello allows me to learn how to write documentation in comment blocks
co-located with code.
}
\author{
\strong{Maintainer}: Karolina Smolinska-Garbulowska \email{karolina.smolinska@icm.uu.se}

}
