% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visunet.R
\name{visunet}
\alias{visunet}
\title{Rule Network Visualistion gadget}
\usage{
visunet(data1, type = "RDF", NewData = FALSE, NewDataValues)
}
\arguments{
\item{rule}{table}
}
\value{
Rule Network Object.
}
\description{
Rule Network Visualistion gadget
}
\examples{

R.ROSETTA output
out=rosetta(autcon)
rules=out$main
visunet(rules)
------------
Line by line file
rules2 = (read.csv2('dataset_ethnicity_all_100set.txt', sep='\\t', header = FALSE, col.names = c('FEATURES', 'DECISION', 'ACC_RHS', 'SUPP_RHS'),stringsAsFactors=FALSE))
rules2$ACC_RHS = as.numeric(rules2$ACC_RHS)
rules2$SUPP_RHS = as.numeric(rules2$SUPP_RHS)
rules2$PVAL = 0.05
d2 = visunet(rules2, 'L')

}
\keyword{misc}
