% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisuNet_pkg.R
\docType{package}
\name{VisuNet}
\alias{VisuNet}
\alias{VisuNet-package}
\title{My first R package: VisuNet}
\description{
A one-line sentence on what this packages does
}
\details{
Single paragraph with more detail on the package.
This can span several lines.
}
