% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visunet.R
\name{visunet}
\alias{visunet}
\title{visunet}
\usage{
visunet(rules, type = "RDF", NodeColorType = "DL", Cust = FALSE,
  CustObject)
}
\arguments{
\item{rules}{data frame or path to the rules file, depending on the values of the type parameter.\cr
In the case of data frame, rule is the output of \code{\link[R.ROSETTA]{rosetta}}. The rules file
is the output of \code{\link[R.ROSETTA]{rosetta}} saved using \code{\link[R.ROSETTA]{saveLineByLine}}
or the output of the ROSETTA GUI version (see \url{http://bioinf.icm.uu.se/rosetta/}).}

\item{type}{character string specifying the type of the input data.  Three types implemented
are "RDF" (the R.ROSETTA data frame), "L" ("Line by line" file format) and "RGUI" (ROSETTA GUI plan format). \cr
The "Line by line" is the tab-separated file, each row corresponds to one rule. File contains 4 or 5 columns:
condition list, decision, accuracy and support of the rule, the optional 5th column contains a fractional support value\cr
The default is "RDF".}

\item{Cust}{logical flag: if TRUE the CustObject is customised Rule Network Object, otherwise the
CustObject parameter is ignored.}

\item{CustObject}{customised Rule Network Object corresponding to the output of the visunet function,
that can be modified according
to the visualisation aim. See more in the details section.}
}
\value{
Rule Network Object.
}
\description{
visunet
}
\examples{
R.ROSETTA output
out = rosetta(autcon)
rules = out$main
vis_out = visunet(rules)
------------
Line by line file
rules = (read.csv2('dataset_ethnicity_all_100set.txt', sep='\\t', header = FALSE, col.names = c('FEATURES', 'DECISION', 'ACC_RHS', 'SUPP_RHS'),stringsAsFactors=FALSE))
rules$ACC_RHS = as.numeric(rules$ACC_RHS)
rules$SUPP_RHS = as.numeric(rules$SUPP_RHS)
rules$PVAL = 0.05
vis_out = visunet(rules, 'L')

}
\keyword{misc}
