% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visunet.R
\name{visunet}
\alias{visunet}
\title{Rule Network Visualistion gadget}
\usage{
visunet(data1, type = "RDF", NewData = FALSE, NewDataValues)
}
\arguments{
\item{rule}{table}
}
\value{
Rule Network Object.
}
\description{
Rule Network Visualistion gadget
}
\examples{
out=rosetta(autcon)
rules=out$main
visunet(rules)
}
\keyword{misc}
